+++ src/main/java/com/abc/account/AbstractAccount.java
@@ -1,8 +1,12 @@
-package com.abc;
+package com.abc.account;
 
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
+
+import com.abc.transaction.DepositTransaction;
+import com.abc.Transaction;
+import com.abc.transaction.WithdrawalTransaction;
 
 public abstract class AbstractAccount implements Account {
