+++ src/main/java/com/abc/transaction/WithdrawalTransaction.java
@@ -1,4 +1,6 @@
-package com.abc;
+package com.abc.transaction;
+
+import com.abc.AbstractTransaction;
 
 public class WithdrawalTransaction extends AbstractTransaction {
