+++ src/test/java/com/abc/account/AccountTest.java
@@ -1,4 +1,4 @@
-package com.abc;
+package com.abc.account;
 
 import org.junit.Before;
 import org.junit.Test;
