+++ src/test/java/com/abc/transaction/TransactionTest.java
@@ -1,5 +1,6 @@
-package com.abc;
+package com.abc.transaction;
 
+import com.abc.Transaction;
 import org.junit.Test;
 
 import static org.hamcrest.CoreMatchers.equalTo;
